\name{clinMMGateway}
\alias{clinMMGateway}
\docType{data}
\title{
Clinical information downloaded from MMRF-Commpass Researcher Gateway 
}
\description{
example of clinical information data.frame downloaded from MMRF-Commpass Researcher Gateway 
}
\usage{clinMMGateway}
\format{
  A data frame with 45 observations on the following 7 variables.
  \describe{
    \item{\code{X}}{a numeric vector}
    \item{\code{public_id}}{a character vector}
    \item{\code{line}}{a numeric vector}
    \item{\code{trtname}}{a character vector}
    \item{\code{trtshnm}}{a character vector}
    \item{\code{bestresp}}{a character vector}
    \item{\code{bestrespsh}}{a character vector}
  }
}

\keyword{datasets}
