% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFutility.R
\name{Convert_toGeneSymbol}
\alias{Convert_toGeneSymbol}
\title{Convert from ensembl.gene to gene.symbol annotation}
\usage{
Convert_toGeneSymbol(ensembl.genes)
}
\arguments{
\item{query}{A query form GDCquery function}
}
\description{
get information about query obteined from GDCquery function
}
\examples{
Convert from ensembl.gene to gene.symbol
ensembl.genes <- c("ENSG00000150676", "ENSG00000099308", "ENSG00000142676", "ENSG00000180776", "ENSG00000108848", "ENSG00000277370", "ENSG00000103811", "ENSG00000101473")
Convert_toGeneSymbol(ensembl.genes)



}
