% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{getSampleSummaryUrlXX}
\alias{getSampleSummaryUrlXX}
\title{Get MMRF-COMPASS Project Summary from GDC}
\usage{
getSampleSummaryUrlXX(project, legacy = FALSE, files.access = NA)

getSampleSummaryUrlXX(project, legacy = FALSE, files.access = NA)
}
\arguments{
\item{project}{is a  GDC project}
}
\description{
Get MMRF-COMPASS Project Summary from GDC

Get MMRF-COMPASS Project Summary from GDC
}
