% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{MMRF_getProjectSummary2}
\alias{MMRF_getProjectSummary2}
\title{Get MMRF-COMPASS Project Summary from GDC}
\usage{
MMRF_getProjectSummary2()
}
\arguments{
\item{project}{is a  GDC project}
}
\description{
Get MMRF-COMPASS Project Summary from GDC
}
\examples{
\dontrun{
getProjectSummary(project="MMRF-COMMPASS")
}
}
