% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFget_InfoCohort}
\alias{MMRFget_InfoCohort}
\title{Get information about the Cohort from GDCquery}
\usage{
MMRFget_InfoCohort(query.mm)
}
\arguments{
\item{query.mm}{is the result of GDCquery function}
}
\value{
the information obout the Cohort obtained from GDCquery function
}
\description{
get information about the Cohort obtained from GDCquery function
}
\examples{
MMRFget_InfoCohort(query.mm)
}
