% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFget_NCasesCohort}
\alias{MMRFget_NCasesCohort}
\title{Get N cases from GDCquery}
\usage{
MMRFget_NCasesCohort(query.mm)
}
\arguments{
\item{query.mm}{is the result of GDCquery function}
}
\value{
the Number of samples in the cohort obtained from GDCquery function
}
\description{
Count the case number obtained from GDCquery function
}
\examples{
MMRFget_NCasesCohort(query.mm)
}
