% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFgetGateway_BOresponse}
\alias{MMRFgetGateway_BOresponse}
\title{Get Best Overall response filtered by sample identifier}
\usage{
MMRFgetGateway_BOresponse(identifier, treat.resp)
}
\arguments{
\item{identifier}{is a vector of samples identifiers}

\item{trt.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into R environment}
}
\value{
a dataframe
}
\description{
filter trt.resp by samples identifiers
}
\examples{
listSamples <- c("MMRF_001","MMRF_002",
                 "MMRF_003","MMRF_003",
                 "MMRF_004","MMRF_005",
                 "MMRF_006","MMRF_007",
                 "MMRF_008","MMRF_009")
                 
bestOveall<-MMRFgetGateway_BOresponse(listSamples, clinMMGateway)              
}
