% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFanalize.R
\name{MMRFGetGateway_BOresponsePlot}
\alias{MMRFGetGateway_BOresponsePlot}
\title{Draw plot of the Best Overall Response to the Treatment}
\usage{
MMRFGetGateway_BOresponsePlot(
  treat.resp,
  therapyname = NULL,
  topN = 20,
  dpi = 100,
  filename = "BestOverall_responsePlot",
  height = 20,
  width = 20
)
}
\arguments{
\item{treat.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into environment}

\item{therapyname}{Therapy name}

\item{topN}{top number of case count}

\item{dpi}{Image dpi}

\item{filename}{The name of the png file}

\item{height}{Image height}

\item{width}{Image width}
}
\value{
table with the case count of the Best overall response to treatments
}
\description{
Draw plot of the Best Overall Response to the Treatment (top 20 to make faster)
}
\examples{
MMRFGetGateway_BOresponsePlot(clinMMGateway,"Bortezomib",height=5, width=8)
MMRFGetGateway_BOresponsePlot(clinMMGateway,topN=40, height=13, width=13)
}
