% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFGetGateway_BOresponseType}
\alias{MMRFGetGateway_BOresponseType}
\title{filter clinical data by Best Overall Response (BOR) type}
\usage{
MMRFGetGateway_BOresponseType(treat.resp, bor)
}
\arguments{
\item{bor}{is the type of Best Overall Response (BOR)
Example:
\tabular{ll}{
CR \tab   Complete Response \cr
PR \tab   Partial Response \cr
VGPR \tab   Very Good Partial Response \cr
SD \tab Stable Disease \cr
PD \tab  Progressive Disease \cr
sCR \tab   Stringent Complete Response \cr
}}

\item{trt.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into R environment}
}
\value{
a dataframe
}
\description{
filter trt.resp by Best Overall Response (BOR) type
}
\examples{
bestOverallType<-MMRFGetGateway_BOresponseType(clinMMGateway,"PR")              
}
