% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFanalize.R
\name{MMRFgetGateway_TimeBestOverallResponsePlot}
\alias{MMRFgetGateway_TimeBestOverallResponsePlot}
\title{Draw plot of Time to Best Overall Response}
\usage{
MMRFgetGateway_TimeBestOverallResponsePlot(
  treat.resp,
  therapyname = NULL,
  ttime = "cycles",
  dpi = 100,
  filename = "TimeBestOverall_responsePlot",
  height = 10,
  width = 10
)
}
\arguments{
\item{treat.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into environment}

\item{therapyname}{Therapy name}

\item{ttime}{cycles/days}

\item{dpi}{Image dpi}

\item{filename}{The name of the png file}

\item{height}{Image height}

\item{width}{Image width}
}
\value{
table with the case count of the Best overall response to treatments
}
\description{
Draw plot of Time to the Best Overall Response
}
\examples{
MMRFgetGateway_TimeBestOverallResponsePlot(clinMMGateway,c("Bortezomib","Dexamethasone"))
MMRFgetGateway_TimeBestOverallResponsePlot(clinMMGateway,c("Bortezomib","Dexamethasone"),"days")
MMRFgetGateway_TimeBestOverallResponsePlot(clinMMGateway,"Bortezomib","days")
}
