% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFanalize.R
\name{MMRFGetGateway_TimeBOresponsePlot}
\alias{MMRFGetGateway_TimeBOresponsePlot}
\title{draws plot correlating the time for the Best Overall Response  (BO) to BO classification}
\usage{
MMRFGetGateway_TimeBOresponsePlot(
  treat.resp,
  therapyname = NULL,
  ttime = "cycles",
  dpi = 100,
  filename = "TimeBestOverall_responsePlot",
  height = 10,
  width = 10
)
}
\arguments{
\item{treat.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into environment}

\item{therapyname}{Therapy name}

\item{ttime}{cycles/days}

\item{dpi}{Image dpi}

\item{filename}{The name of the png file}

\item{height}{Image height}

\item{width}{Image width}
}
\value{
table with the case count of the Best overall response to treatments
}
\description{
Draw plot of Time to the Best Overall Response
}
\examples{
MMRFGetGateway_TimeBOresponsePlot(clinMMGateway,c("Bortezomib","Dexamethasone"))
MMRFGetGateway_TimeBOresponsePlot(clinMMGateway,c("Bortezomib","Dexamethasone"),"days")
MMRFGetGateway_TimeBOresponsePlot(clinMMGateway,"Bortezomib","days")
}
