% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFanalize.R
\name{MMRFGetGateway_TrtBOdurationPlot}
\alias{MMRFGetGateway_TrtBOdurationPlot}
\title{Draw plot of Treatment duration filtered by Therapy classification}
\usage{
MMRFGetGateway_TrtBOdurationPlot(
  treat.resp,
  therapyname = NULL,
  ttime = "cycles",
  line = 1,
  bor = "CR",
  dpi = 100,
  filename = "Trt_DurationPlot",
  height = 8,
  width = 8
)
}
\arguments{
\item{treat.resp}{is a data.frame of clinical information downloaded from MMRF-Commpass Researcher Gateway 
and imported into environment}

\item{therapyname}{Therapy name}

\item{ttime}{cycles/days}

\item{line}{Line of therapy}

\item{bor}{is the type of BO Response 
Example:
\tabular{ll}{
CR \tab   Complete Response \cr
PR \tab   Partial Response \cr
VGPR \tab   Very Good Partial Response \cr
SD \tab Stable Disease \cr
PD \tab  Progressive Disease \cr
sCR \tab   Stringent Complete Response \cr
}}

\item{dpi}{Image dpi}

\item{filename}{The name of the png file}

\item{height}{Image height}

\item{width}{Image width}
}
\description{
Draw plot of the Treatment duration cycle or days
}
\examples{
MMRFGetGateway_TrtBOdurationPlot(clinMMGateway,"Bortezomib",ttime="cycles",bor="PR",height=10, width=10)
MMRFGetGateway_TrtBOdurationPlot(clinMMGateway,"Bortezomib",ttime="days",bor="VGPR",height=10, width=10)
MMRFGetGateway_TrtBOdurationPlot(clinMMGateway,c("Bortezomib","Lenalidomide"),ttime="days",bor="VGPR",height=10, width=10)
}
