% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFGetGDC_IdentifierByTherapy}
\alias{MMRFGetGDC_IdentifierByTherapy}
\title{Search patient clinical information filtered by therapy name}
\usage{
MMRFGetGDC_IdentifierByTherapy(therapyname, clin.mm)
}
\arguments{
\item{therapyname}{Therapy name}

\item{clin.mm}{is a data.frame using function 'clinic' with information
related to identifier / samples such as bcr_patient_barcode, days_to_death ,
days_to_last_follow_up , vital_status, etc}
}
\value{
a character vector with samples identifiers
}
\description{
Search patient clinical information filtered by therapy name
}
\examples{
bar.dexa<-MMRFGetGDC_BarcodeTherapy("Dexamethasone",clin.mm)
}
