% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFgetGDC_Treatments}
\alias{MMRFgetGDC_Treatments}
\title{Get list of treatments}
\usage{
MMRFgetGDC_Treatments(clin.mm)
}
\arguments{
\item{clin.mm}{is a data.frame using function 'clinic' with information
related to barcode / samples such as bcr_patient_barcode, days_to_death ,
days_to_last_follow_up , vital_status, etc}
}
\value{
a data.frame
}
\description{
get all treatmentsSearch patient clinical information filtered by therapy name
}
