% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFqueryGDC_clinic}
\alias{MMRFqueryGDC_clinic}
\title{Get GDC clinical data}
\usage{
MMRFqueryGDC_clinic(type = "clinical", save.csv = FALSE)
}
\arguments{
\item{type}{A valid type. Options "clinical", "Biospecimen"  (see list with getGDCprojects()$project_id)]}

\item{save.csv}{Write clinical information into a csv document}
}
\value{
A data frame with the clinical information
}
\description{
MMRFqueryGDC_clinic will download and prepare all clinical information from the API
as the one with using the button from each project
}
\examples{
clin.mm<-MMRFqueryGDC_clinic(type = "clinical")
clin.mm<-MMRFqueryGDC_clinic(type = "Biospecimen")
}
