% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFutility.R
\name{MMRFqueryGDC_Summary}
\alias{MMRFqueryGDC_Summary}
\title{Get GDC Query Summary}
\usage{
MMRFqueryGDC_Summary(query)
}
\arguments{
\item{query}{A query form GDCquery function}
}
\value{
a data.frame
}
\description{
get information about query obteined from GDCquery function
}
\examples{
\dontrun{
query.mm.fpkm <- GDCquery(project = "MMRF-COMMPASS",
                          data.category = "Transcriptome Profiling",
                          data.type = "Gene Expression Quantification",
                          workflow.type="HTSeq - FPKM")
summary<-MMRFqueryGDC_Summary(query.mm.fpkm)


}
}
