% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMRFclinical.R
\name{MMRFGetGDC_Treatments}
\alias{MMRFGetGDC_Treatments}
\title{Get list of treatments}
\usage{
MMRFGetGDC_Treatments(clin.mm)
}
\arguments{
\item{clin.mm}{is a data.frame using function 'clinic' with information
related to barcode / samples such as bcr_patient_barcode, days_to_death ,
days_to_last_follow_up , vital_status, etc}
}
\value{
a data.frame
}
\description{
get treatment list from clinical data
}
\examples{
treat.list<-MMRFGetGDC_Treatments(clin.mm)
}
